Creación de test de modelo.

+++ src/main/java/ar/edu/unq/eperdemic/controller/dto/CoordenadasDTO.kt
@@ -1,10 +1,9 @@
 package ar.edu.unq.eperdemic.controller.dto
 
 import ar.edu.unq.eperdemic.modelo.Coordenada
-import ar.edu.unq.eperdemic.modelo.Especie
 
-class CoordenadasDTO ( val latitud: Int,
-                       val longitud: Int ) {
+class CoordenadasDTO ( val latitud: Double,
+                       val longitud: Double ) {
 
     fun aModelo(): Coordenada {
         return Coordenada(latitud, longitud)
